Renamed middlewares to middleware
